build: Fix the build (pin to older setuptools; add missing asn1crypto)
In the newer environment, setuptools 46 got included. This one doesn't
have the Feature class anymore, and pyScss installer tripped on it.
Downgrade setuptools.

Additionally, the apt-get autoremove after uninstalling python3-pip
would now also remove some more crypto packages, including asn1crypto,
which was required by glance. That cause the following strange error:

  2 is not a supported API version for the "image" service,
  choices are: 2

This was caused the failing of BOTH these:

  VERSIONS.load_supported_version(2, {"client": glance_client_v2,
                                      "version": 2})

  VERSIONS.load_supported_version(1, {"client": glance_client_v1,
                                      "version": 1})

That caused only a a stub (preferred) {2: {"version": 2}} to get inserted.
While the rest of the code expected a:
  {Version('2'): {"version": 2, "client": <Client>}}